admin wallet page european number system fix

+++ homepage/models.py
@@ -6,7 +6,7 @@ from decimal import Decimal
 class User(BaseUser):
     name = models.CharField(max_length=100, default="")
     secretcode = models.CharField(max_length=100,default="")
-    coins = models.IntegerField(default=0)
+    coins = models.DecimalField(decimal_places=2,max_digits=10,default=0)
     estate = models.DecimalField(decimal_places=2,max_digits=10,default=0)
     
     updated_at = models.DateTimeField(auto_now=True)
